Remove legacy implementations
