--- intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/problems/PluginProblemsLoaderFactory.kt
@@ -1,11 +1,0 @@
-package com.jetbrains.plugin.structure.intellij.problems
-
-fun interface PluginProblemsLoaderFactory {
-  fun getPluginProblemsLoader(): PluginProblemsLoader
-}
-
-object ClassPathPluginProblemsLoaderFactory: PluginProblemsLoaderFactory {
-  override fun getPluginProblemsLoader(): PluginProblemsLoader {
-    return PluginProblemsLoader.fromClassPath()
-  }
-}
