+++ intellij-plugin-verifier/verifier-cli/src/main/java/com/jetbrains/pluginverifier/options/PluginParsingConfigurationResolution.kt
@@ -30,30 +30,4 @@ class PluginParsingConfigurationResolution {
       defaultResolver
     }
   }
-
-  fun resolveLevelRemapping(configuration: PluginParsingConfiguration,
-                            pluginProblemsLoaderFactory: PluginProblemsLoaderFactory
-                            = ClassPathPluginProblemsLoaderFactory): PluginCreationResultResolver {
-    val defaultResolver = IntelliJPluginCreationResultResolver()
-    return if (configuration.pluginSubmissionType == SubmissionType.EXISTING) {
-      val problemLevelRemapping = try {
-        val pluginProblemsLoader = pluginProblemsLoaderFactory.getPluginProblemsLoader()
-        val problemLevelRemappingDefinitions = pluginProblemsLoader.load()
-        val pluginProblemSet = problemLevelRemappingDefinitions["existing-plugin"]
-          ?: emptyProblemLevelRemapping("existing-plugin").also {
-            LOG.warn(("Plugin problem remapping definition 'existing-plugin' was not found. Problems will not be remapped"))
-          }
-        val parser = RemappedPluginProblemLevelParser()
-        val remapping = parser.parse(pluginProblemSet)
-        remapping
-      } catch (e: IOException) {
-        LOG.error(e.message, e)
-        emptyMap()
-      }
-
-      LevelRemappingPluginCreationResultResolver(defaultResolver, additionalLevelRemapping = problemLevelRemapping)
-    } else {
-      defaultResolver
-    }
-  }
 }
